feat: move config load-up -> Context:try_build
