+++ cli/src/cli.rs
@@ -9,7 +9,8 @@ use std::path::PathBuf;
 use std::sync::Arc;
 
 use crate::keypair::{get_keypair_path, get_payer};
-use crate::config::TapeConfig;
+use crate::config::{TapeConfig, TapeConfigError};
+use crate::log;
 
 #[derive(Parser)]
 #[command(
@@ -239,7 +240,68 @@ pub struct Context {
 }
 
 impl Context{
-    pub fn try_build(_cli:&Cli, config: &TapeConfig) -> Result<Self> {
+    pub fn try_build(cli:&Cli) -> Result<Self> {
+        
+        // loading up configs
+        let config = match TapeConfig::load(&cli.config) {
+            Ok(config) => config,
+            Err(e) => match e {
+                TapeConfigError::ConfigFileNotFound => {
+                    log::print_info("tape.toml not found, creating default configuration...");
+                    match TapeConfig::create_default() {
+                        Ok(config) => {
+                            log::print_info("✓ Default configuration created successfully");
+                            config
+                        },
+                        Err(creation_error) => {
+                            log::print_error(&format!("{}", creation_error));
+                            std::process::exit(1);
+                        }
+                    }
+                },
+
+                TapeConfigError::CustomConfigFileNotFound(path) => {
+                // This happens when user explicitly provided a path that doesn't exist
+                log::print_error(&format!("Custom config file not found: {}", path));
+                log::print_info("Please check the path and try again.");
+                std::process::exit(1);
+            },
+                
+                TapeConfigError::InvalidUrl(msg) => {
+                    log::print_error(&format!("URL Configuration Error: {}", msg));
+                    log::print_info("Please fix the URL in your tape.toml file and try again.");
+                    std::process::exit(1);
+                },
+                
+                TapeConfigError::KeypairNotFound(path) => {
+                    log::print_error(&format!("Keypair not found at path: {}", path));
+                    log::print_info("Please ensure the keypair file exists at the specified path in tape.toml");
+                    std::process::exit(1);
+                },
+                
+                TapeConfigError::FileReadError(io_err) => {
+                    log::print_error(&format!("Could not read config file: {}", io_err));
+                    std::process::exit(1);
+                },
+                
+                TapeConfigError::ParseError(parse_err) => {
+                    log::print_error(&format!("Invalid tape.toml format: {}", parse_err));
+                    log::print_info("Please check your tape.toml file syntax.");
+                    std::process::exit(1);
+                },
+                
+                TapeConfigError::HomeDirectoryNotFound => {
+                    log::print_error("Could not determine home directory");
+                    std::process::exit(1);
+                },
+                
+                TapeConfigError::DefaultConfigCreationFailed(msg) => {
+                    log::print_error(&format!("Failed to create default config: {}", msg));
+                    std::process::exit(1);
+                },
+            }
+        };
+
         let rpc_url = config.solana.rpc_url.to_string();
         let commitment_level = config.solana.commitment.to_commitment_config();
         let rpc = Arc::new(
