+++ cli/src/main.rs
@@ -14,7 +14,6 @@ use env_logger::{self, Env};
 use tape_network::store::TapeStore;
 
 use crate::cli::Context;
-use crate::config::{TapeConfig, TapeConfigError};
 
 
 fn main() -> Result<()>{
@@ -40,66 +39,7 @@ async fn run_tape_cli() -> Result<()> {
 
     let cli = Cli::parse();
 
-    let config = match TapeConfig::load(&cli.config) {
-        Ok(config) => config,
-        Err(e) => match e {
-            TapeConfigError::ConfigFileNotFound => {
-                log::print_info("tape.toml not found, creating default configuration...");
-                match TapeConfig::create_default() {
-                    Ok(config) => {
-                        log::print_info("✓ Default configuration created successfully");
-                        config
-                    },
-                    Err(creation_error) => {
-                        log::print_error(&format!("{}", creation_error));
-                        std::process::exit(1);
-                    }
-                }
-            },
-
-            TapeConfigError::CustomConfigFileNotFound(path) => {
-            // This happens when user explicitly provided a path that doesn't exist
-            log::print_error(&format!("Custom config file not found: {}", path));
-            log::print_info("Please check the path and try again.");
-            std::process::exit(1);
-        },
-            
-            TapeConfigError::InvalidUrl(msg) => {
-                log::print_error(&format!("URL Configuration Error: {}", msg));
-                log::print_info("Please fix the URL in your tape.toml file and try again.");
-                std::process::exit(1);
-            },
-            
-            TapeConfigError::KeypairNotFound(path) => {
-                log::print_error(&format!("Keypair not found at path: {}", path));
-                log::print_info("Please ensure the keypair file exists at the specified path in tape.toml");
-                std::process::exit(1);
-            },
-            
-            TapeConfigError::FileReadError(io_err) => {
-                log::print_error(&format!("Could not read config file: {}", io_err));
-                std::process::exit(1);
-            },
-            
-            TapeConfigError::ParseError(parse_err) => {
-                log::print_error(&format!("Invalid tape.toml format: {}", parse_err));
-                log::print_info("Please check your tape.toml file syntax.");
-                std::process::exit(1);
-            },
-            
-            TapeConfigError::HomeDirectoryNotFound => {
-                log::print_error("Could not determine home directory");
-                std::process::exit(1);
-            },
-            
-            TapeConfigError::DefaultConfigCreationFailed(msg) => {
-                log::print_error(&format!("Failed to create default config: {}", msg));
-                std::process::exit(1);
-            },
-        }
-    };
-
-    let context = Context::try_build(&cli, &config)?;
+    let context = Context::try_build(&cli)?;
 
     match cli.command {
         Commands::Init {} |
